Restructured the project for visibility.
