+++ src/main/kotlin/com/appdetex/devops/Application.kt
@@ -1,58 +1,26 @@
 package com.appdetex.devops
 
+import com.appdetex.devops.config.appConfig
+import com.appdetex.devops.controllers.ecsTestController
+import com.appdetex.devops.controllers.renderTestController
+import com.appdetex.devops.controllers.sessionController
 import com.appdetex.devops.domain.EnvironmentVersionsDTO
 import com.appdetex.devops.domain.Project
-import com.appdetex.devops.service.EcsService
-import com.fasterxml.jackson.databind.SerializationFeature
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import freemarker.cache.ClassTemplateLoader
+import com.appdetex.devops.exception.AuthenticationException
+import com.appdetex.devops.exception.AuthorizationException
 import io.ktor.application.Application
-import io.ktor.application.ApplicationCall
 import io.ktor.application.call
 import io.ktor.application.install
-import io.ktor.features.AutoHeadResponse
-import io.ktor.features.CORS
-import io.ktor.features.CallLogging
-import io.ktor.features.ConditionalHeaders
-import io.ktor.features.ContentNegotiation
 import io.ktor.features.StatusPages
-import io.ktor.freemarker.FreeMarker
-import io.ktor.freemarker.FreeMarkerContent
-import io.ktor.html.respondHtml
-import io.ktor.http.ContentType
-import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
-import io.ktor.http.content.resources
-import io.ktor.http.content.static
-import io.ktor.jackson.jackson
-import io.ktor.request.path
 import io.ktor.response.respond
 import io.ktor.response.respondText
 import io.ktor.routing.get
 import io.ktor.routing.routing
-import io.ktor.sessions.Sessions
-import io.ktor.sessions.cookie
-import io.ktor.sessions.get
-import io.ktor.sessions.sessions
-import io.ktor.sessions.set
-import kotlinx.css.CSSBuilder
-import kotlinx.css.Color
-import kotlinx.css.body
-import kotlinx.css.em
-import kotlinx.css.p
-import kotlinx.html.CommonAttributeGroupFacade
-import kotlinx.html.FlowOrMetaDataContent
-import kotlinx.html.body
-import kotlinx.html.h1
-import kotlinx.html.li
-import kotlinx.html.style
-import kotlinx.html.ul
 import org.koin.core.Koin
-import org.koin.ktor.ext.inject
 import org.koin.log.PrintLogger
 import org.koin.standalone.StandAloneContext.startKoin
-import org.slf4j.event.Level
+
 
 fun main(args: Array<String>) {
     Koin.logger = PrintLogger()
@@ -66,56 +34,16 @@ fun main(args: Array<String>) {
 @kotlin.jvm.JvmOverloads
 fun Application.module(testing: Boolean = false) {
 
-    val ecsService: EcsService by inject()
+    // Configuration
+    appConfig()
 
-    install(FreeMarker) {
-        templateLoader = ClassTemplateLoader(this::class.java.classLoader, "templates")
-    }
-
-    install(Sessions) {
-        cookie<MySession>("MY_SESSION") {
-            cookie.extensions["SameSite"] = "lax"
-        }
-    }
-
-    install(AutoHeadResponse)
-
-    install(CallLogging) {
-        level = Level.INFO
-        filter { call -> call.request.path().startsWith("/") }
-    }
-
-    install(ConditionalHeaders)
-
-    install(CORS) {
-        method(HttpMethod.Options)
-        method(HttpMethod.Put)
-        method(HttpMethod.Delete)
-        method(HttpMethod.Patch)
-        header(HttpHeaders.Authorization)
-        header("MyCustomHeader")
-        allowCredentials = true
-        anyHost() // TODO: Don't do this in production if possible. Try to limit it.
-    }
-
-    install(ContentNegotiation) {
-        jackson {
-            enable(SerializationFeature.INDENT_OUTPUT)
-            registerModule(JavaTimeModule())
-        }
-    }
-
+    // Routes
     routing {
-        get("/") {
 
-            val response = ecsService.getTasks("Waits")
-
-            val responseHeader = "FAMILY - STATUS - IMAGE"
-            val responseText = response.values.sortedBy { it.family() }
-                .joinToString(separator = "\n") { "${it.family()} - ${it.statusAsString()} - ${it.containerDefinitions()[0].image()}"}
-
-            call.respondText("$responseHeader\n$responseText", contentType = ContentType.Text.Plain)
-        }
+        // Controllers
+        ecsTestController()
+        renderTestController()
+        sessionController()
 
         get("/versions/{environment}") {
             call.parameters["environment"]?.takeIf { env -> env == "test" }?.let { environment ->
@@ -123,81 +51,11 @@ fun Application.module(testing: Boolean = false) {
             } ?: call.respondText(text = "Invalid environment name='${call.parameters["environment"]}'", status = HttpStatusCode.BadRequest)
         }
 
-        get("/html-dsl") {
-            call.respondHtml {
-                body {
-                    h1 { +"HTML" }
-                    ul {
-                        for (n in 1..10) {
-                            li { +"$n" }
-                        }
-                    }
-                }
-            }
-        }
-
-        get("/styles.css") {
-            call.respondCss {
-                body {
-                    backgroundColor = Color.red
-                }
-                p {
-                    fontSize = 2.em
-                }
-                rule("p.myclass") {
-                    color = Color.blue
-                }
-            }
-        }
-
-        get("/html-freemarker") {
-            call.respond(FreeMarkerContent("index.ftl", mapOf("data" to IndexData(listOf(1, 2, 3))), ""))
-        }
-
-        // Static feature. Try to access `/static/ktor_logo.svg`
-        static("/static") {
-            resources("static")
-        }
-
-        get("/session/increment") {
-            val session = call.sessions.get<MySession>() ?: MySession()
-            call.sessions.set(session.copy(count = session.count + 1))
-            call.respondText("Counter is ${session.count}. Refresh to increment.")
-        }
-
+        // Status Page Definitions
         install(StatusPages) {
-            exception<AuthenticationException> { cause ->
-                call.respond(HttpStatusCode.Unauthorized)
-            }
-            exception<AuthorizationException> { cause ->
-                call.respond(HttpStatusCode.Forbidden)
-            }
-
-        }
-
-        get("/json/jackson") {
-            call.respond(mapOf("hello" to "world"))
+            exception<AuthenticationException> { call.respond(HttpStatusCode.Unauthorized) }
+            exception<AuthorizationException> { call.respond(HttpStatusCode.Forbidden) }
         }
     }
 }
 
-data class IndexData(val items: List<Int>)
-
-data class MySession(val count: Int = 0)
-
-class AuthenticationException : RuntimeException()
-class AuthorizationException : RuntimeException()
-
-fun FlowOrMetaDataContent.styleCss(builder: CSSBuilder.() -> Unit) {
-    style(type = ContentType.Text.CSS.toString()) {
-        +CSSBuilder().apply(builder).toString()
-    }
-}
-
-fun CommonAttributeGroupFacade.style(builder: CSSBuilder.() -> Unit) {
-    this.style = CSSBuilder().apply(builder).toString().trim()
-}
-
-suspend inline fun ApplicationCall.respondCss(builder: CSSBuilder.() -> Unit) {
-    this.respondText(CSSBuilder().apply(builder).toString(), ContentType.Text.CSS)
-}
