+++ src/main/kotlin/com/appdetex/devops/config/AppConfig.kt
@@ -1,0 +1,59 @@
+package com.appdetex.devops.config
+
+import com.appdetex.devops.domain.MySession
+import com.fasterxml.jackson.databind.SerializationFeature
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import com.fasterxml.jackson.module.kotlin.KotlinModule
+import freemarker.cache.ClassTemplateLoader
+import io.ktor.application.Application
+import io.ktor.application.install
+import io.ktor.features.*
+import io.ktor.freemarker.FreeMarker
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.jackson.jackson
+import io.ktor.request.path
+import io.ktor.sessions.Sessions
+import io.ktor.sessions.cookie
+import org.slf4j.event.Level
+
+fun Application.appConfig() {
+
+    install(FreeMarker) {
+        templateLoader = ClassTemplateLoader(this::class.java.classLoader, "templates")
+    }
+
+    install(Sessions) {
+        cookie<MySession>("MY_SESSION") {
+            cookie.extensions["SameSite"] = "lax"
+        }
+    }
+
+    install(CallLogging) {
+        level = Level.INFO
+        filter { call -> call.request.path().startsWith("/") }
+    }
+
+    install(CORS) {
+        method(HttpMethod.Options)
+        method(HttpMethod.Put)
+        method(HttpMethod.Delete)
+        method(HttpMethod.Patch)
+        header(HttpHeaders.Authorization)
+        header("MyCustomHeader")
+        allowCredentials = true
+        anyHost() // TODO: Don't do this in production if possible. Try to limit it.
+    }
+
+    install(ConditionalHeaders)
+    install(AutoHeadResponse)
+
+    install(ContentNegotiation) {
+        jackson {
+            enable(SerializationFeature.INDENT_OUTPUT)
+            registerModule(JavaTimeModule())
+            registerModule(KotlinModule())
+        }
+    }
+
+}
