+++ src/main/kotlin/com/appdetex/devops/domain/IndexData.kt
@@ -1,0 +1,3 @@
+package com.appdetex.devops.domain
+
+data class IndexData(val items: List<Int>)
