+++ src/main/kotlin/com/appdetex/devops/domain/SessionData.kt
@@ -1,0 +1,3 @@
+package com.appdetex.devops.domain
+
+data class MySession(val count: Int = 0)
