+++ src/main/kotlin/com/appdetex/devops/exception/AuthenticationException.kt
@@ -1,0 +1,4 @@
+package com.appdetex.devops.exception
+
+
+class AuthenticationException : RuntimeException()
