+++ src/main/kotlin/com/appdetex/devops/exception/AuthorizationException.kt
@@ -1,0 +1,4 @@
+package com.appdetex.devops.exception
+
+
+class AuthorizationException : RuntimeException()
